+++ src/main/java/kr/co/monitoringserver/controller/api/AdminApiController.java
@@ -1,27 +1,25 @@
 package kr.co.monitoringserver.controller.api;
 
 import jakarta.validation.Valid;
+import kr.co.monitoringserver.persistence.repository.UserAttendanceRepository;
 import kr.co.monitoringserver.service.dtos.request.AdminReqDTO;
 import kr.co.monitoringserver.service.dtos.response.ResDTO;
 import kr.co.monitoringserver.service.service.AdminService;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.domain.Pageable;
-import org.springframework.data.domain.Sort;
-import org.springframework.data.web.PageableDefault;
 import org.springframework.http.HttpStatus;
-import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
-import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 
-import java.time.LocalDate;
 import java.util.Map;
 @RestController
 public class AdminApiController {
     @Autowired
     private AdminService adminService;
+
+    @Autowired
+    private UserAttendanceRepository userAttendanceRepository;
 
     /**
      * updateUser : 회원정보를 Update 한다.
